requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
google-generativeai==0.8.5
gtts==2.5.4
deep-translator==1.11.4
scikit-learn==1.4.0
numpy==1.26.4
python-dotenv==1.0.1
mangum==0.17.0
